terminal_chat/requirements.txt
openai
langchain
python-dotenv
argparse
